sql/jobs: move package out of sql
Jobs have nothing to do with SQL, except that they use the SQL interface
to query and manipulate the jobs in the cluster, so move the sql/jobs
package out of the sql directory. Jobs are important enough as a
standalone entity to warrant this.

Release note: None
