+++ pkg/jobs/registry_test.go
